add unit and e2e test
