+++ src/components/NavBar.astro
@@ -21,7 +21,7 @@ const isAlbumsActive = currentPath.startsWith('/discography')
       </div>
 
       <!-- Center logo -->
-      <a href="/" class="group relative mx-8 flex items-center justify-center">
+      <a href="/" class="group relative mx-8 flex items-center justify-center" aria-label="Home">
         <div class="relative h-12 w-12 md:h-14 md:w-14">
           <Image
             class="relative transition-transform duration-300 group-hover:scale-105"
