+++ src/const/discography.ts
@@ -1,5 +1,5 @@
 import type { TDiscography } from '../types/discography'
-import { Unit } from './members'
+import { DISCOGRAPHY_DATA } from './discography-data'
 import tenTheStoryGoesOn from '../assets/discography/ten-the-story-goes-on.webp'
 import blink from '../assets/discography/blink.webp'
 import lilFantasyVol1 from '../assets/discography/lil-fantasy-vol-1.webp'
@@ -95,1918 +95,105 @@ import aStrangeDay from '../assets/discography/a-strange-day.webp'
 import popStar from '../assets/discography/pop-star.webp'
 import youAreTheAppleOfMyEye from '../assets/discography/you-are-the-apple-of-my-eye.webp'
 
-export const DISCOGRAPHY: TDiscography[] = [
-  {
-    id: 'sotsugyou-cover',
-    name: 'Sotsugyou Cover (Sana)',
-    image: sotsugyouCover,
-    unit: [Unit.sana],
-    releaseDate: new Date('2021-02-24'),
-    types: ['single'],
-    market: 'japanese',
-    songs: [{ title: '卒業 (Sotsugyou)' }],
-  },
-  {
-    id: 'like-a-star',
-    name: 'Like a Star',
-    image: likeaStar,
-    unit: [Unit.jeongyeon],
-    releaseDate: new Date('2018-10-25'),
-    types: ['ost'],
-    market: 'korean',
-    songs: [{ title: 'Like a Star' }],
-  },
-  {
-    id: 'i-fly',
-    name: 'I FLY',
-    image: iFly,
-    unit: [Unit.jihyo],
-    releaseDate: new Date('2022-07-29'),
-    types: ['ost'],
-    market: 'korean',
-    songs: [{ title: 'I FLY' }],
-  },
-  {
-    id: 'a-strange-day',
-    name: 'A Strange Day',
-    image: aStrangeDay,
-    unit: [Unit.jihyo],
-    releaseDate: new Date('2022-11-14'),
-    types: ['ost'],
-    market: 'korean',
-    songs: [{ title: 'A Strange Day' }],
-  },
-  {
-    id: 'popstar-remix',
-    name: 'Pop Star (CHAEYOUNG Remix)',
-    image: popStar,
-    unit: [Unit.chaeyoung],
-    releaseDate: new Date('2023-10-25'),
-    types: ['single'],
-    market: 'english',
-    songs: [{ title: 'Pop Star (CHAEYOUNG Remix)' }],
-  },
-  {
-    id: 'you-are-the-apple-of-my-eye',
-    name: 'You Are the Apple of My Eye',
-    image: youAreTheAppleOfMyEye,
-    unit: [Unit.dahyun],
-    releaseDate: new Date('2025-02-20'),
-    types: ['ost'],
-    market: 'korean',
-    songs: [{ title: 'You Are the Apple of My Eye' }],
-  },
-  {
-    id: 'the-story-begins',
-    name: 'The Story Begins',
-    image: theStoryBegins,
-    unit: [Unit.twice],
-    releaseDate: new Date('2015-10-20'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'Like OOH-AHH',
-        link: 'https://youtu.be/0rtV5esQT6I?si=XDDHfSUtX8-HyuiB',
-      },
-      {
-        title: 'Do It Again',
-      },
-      {
-        title: 'Going Crazy',
-      },
-      {
-        title: 'Truth',
-      },
-      {
-        title: 'Candy Boy',
-      },
-      {
-        title: 'Like a Fool',
-      },
-    ],
-  },
-  {
-    id: 'page-two',
-    name: 'Page Two',
-    image: pageTwo,
-    unit: [Unit.twice],
-    releaseDate: new Date('2016-04-25'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'CHEER UP',
-        link: 'https://youtu.be/c7rCyll5AeY?si=ynE2G4Pbk6wUMkiG',
-      },
-      {
-        title: 'Precious Love',
-      },
-      {
-        title: 'Touchdown',
-      },
-      {
-        title: 'Tuk Tok',
-      },
-      {
-        title: 'Woohoo',
-      },
-      {
-        title: 'My Headphones On',
-      },
-      {
-        title: "I'm Going To Be a Star",
-      },
-    ],
-  },
-  {
-    id: 'twicecoaster-lane-1',
-    name: 'TWICEcoaster: Lane 1',
-    image: twicecoasterLane1,
-    unit: [Unit.twice],
-    releaseDate: new Date('2016-10-24'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'TT',
-        link: 'https://youtu.be/ePpPVE-GGJw?si=rKbIn6YxCFZi8Pwr',
-      },
-      {
-        title: '1 to 10',
-      },
-      {
-        title: 'Ponytail',
-      },
-      {
-        title: 'Jelly Jelly',
-      },
-      {
-        title: 'Pit-A-Pat',
-      },
-      {
-        title: 'Next Page',
-      },
-      {
-        title: 'One in a Million',
-      },
-    ],
-  },
-  {
-    id: 'twicecoaster-lane-2',
-    name: 'TWICEcoaster: Lane 2',
-    image: twicecoasterLane2,
-    unit: [Unit.twice],
-    releaseDate: new Date('2017-02-20'),
-    types: ['repackage', 'mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'Knock Knock',
-        link: 'https://youtu.be/8A2t_tAjMz8?si=0FU4jnOL6hZLVYr9',
-      },
-      {
-        title: 'Ice Cream',
-      },
-      {
-        title: 'TT',
-      },
-      {
-        title: '1 to 10',
-      },
-      {
-        title: 'Ponytail',
-      },
-      {
-        title: 'Jelly Jelly',
-      },
-      {
-        title: 'Pit-A-Pat',
-      },
-      {
-        title: 'Next Page',
-      },
-      {
-        title: 'One in a Million',
-      },
-      {
-        title: 'TT (Tak Remix)',
-      },
-      {
-        title: 'Like Ooh-Ahh (Instrumental)',
-      },
-      {
-        title: 'TT (Instrumental)',
-      },
-      {
-        title: 'Cheer Up (Instrumental)',
-      },
-      {
-        title: 'TT (Instrumental)',
-      },
-    ],
-  },
-  {
-    id: 'whats-this',
-    name: "Whats's This?",
-    image: whatsThis,
-    unit: [Unit.twice],
-    releaseDate: new Date('2017-02-24'),
-    types: ['mini-album', 'digital'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'Like Ooh-Ahh',
-      },
-      {
-        title: 'Cheer Up',
-      },
-      {
-        title: 'Touchdown',
-      },
-    ],
-  },
-  {
-    id: 'signal',
-    name: 'Signal',
-    image: signal,
-    unit: [Unit.twice],
-    releaseDate: new Date('2017-05-15'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'Signal',
-        link: 'https://youtu.be/VQtonf1fv_s?si=8c9ZkxxgJNs8w-MF',
-      },
-      {
-        title: 'Three Times a Day',
-      },
-      {
-        title: 'Only You',
-      },
-      {
-        title: 'Hold Me Tight',
-      },
-      {
-        title: 'Eye Eye Eyes',
-      },
-      {
-        title: 'Someone Like Me',
-      },
-    ],
-  },
-  {
-    id: 'hashtag-twice',
-    name: '#TWICE',
-    image: hashtagTwice,
-    unit: [Unit.twice],
-    types: ['mini-album'],
-    market: 'japanese',
-    releaseDate: new Date('2017-06-28'),
-    songs: [
-      {
-        title: 'Like Ooh-Ahh (Japanese Ver.)',
-        link: 'https://youtu.be/1iLF3zr5mUM?si=hKy28HCNWi6eWTKo',
-      },
-      {
-        title: 'Cheer Up (Japanese Ver.)',
-      },
-      {
-        title: 'TT (Japanese Ver.)',
-        link: 'https://youtu.be/t35H2BVq490?si=WZMGBdoeRk1BlOSN',
-      },
-      {
-        title: 'Knock Knock (Japanese Ver.)',
-      },
-      {
-        title: 'Signal (Japanese Ver.)',
-        link: 'https://youtu.be/nrNWpxFflr4?si=n4hIefOxgb40ZZNg',
-      },
-      {
-        title: 'Like Ooh-Ahh',
-      },
-      {
-        title: 'Cheer Up',
-      },
-      {
-        title: 'TT',
-      },
-      {
-        title: 'Knock Knock',
-      },
-      {
-        title: 'Signal',
-      },
-    ],
-  },
-  {
-    id: 'one-more-time',
-    name: 'One More Time',
-    image: oneMoreTime,
-    unit: [Unit.twice],
-    types: ['single'],
-    market: 'japanese',
-    releaseDate: new Date('2017-10-18'),
-    songs: [
-      {
-        title: 'One More Time',
-        link: 'https://youtu.be/HuoOEry-Yc4?si=bw4DgoEq7dBJdtQ4',
-      },
-      {
-        title: 'Luv Me',
-      },
-      {
-        title: 'One More Time (Instrumental)',
-      },
-      {
-        title: 'Luv Me (Instrumental)',
-      },
-    ],
-  },
-  {
-    id: 'twicetagram',
-    name: 'Twicetagram',
-    image: twicetagram,
-    unit: [Unit.twice],
-    types: ['full-album'],
-    market: 'korean',
-    releaseDate: new Date('2017-10-30'),
-    songs: [
-      {
-        title: 'Likey',
-        link: 'https://youtu.be/V2hlQkVJZhE?si=I0yxXN-kiY95c1Qm',
-      },
-      {
-        title: 'Turtle',
-      },
-      {
-        title: 'Missing U',
-      },
-      {
-        title: 'Wow',
-      },
-      {
-        title: 'FFW',
-      },
-      {
-        title: 'Ding Dong',
-      },
-      {
-        title: '24/7',
-      },
-      {
-        title: 'Look At Me',
-      },
-      {
-        title: 'Rollin',
-      },
-      {
-        title: 'Love Line',
-      },
-      {
-        title: "Don't Give Up",
-      },
-      {
-        title: 'You In My Heart',
-      },
-      {
-        title: 'Jaljayo Good Night',
-      },
-    ],
-  },
-  {
-    id: 'merry-and-happy',
-    name: 'Merry & Happy',
-    image: merryAndHappy,
-    unit: [Unit.twice],
-    types: ['repackage', 'full-album'],
-    market: 'korean',
-    releaseDate: new Date('2017-12-11'),
-    songs: [
-      {
-        title: 'Heart Shaker',
-        link: 'https://youtu.be/rRzxEiBLQCA?si=TC6Z6HoAXJndc2TO',
-      },
-      {
-        title: 'Merry & Happy',
-        link: 'https://youtu.be/zi_6oaQyckM?si=9Ai4GdGz6GmITJ2I',
-      },
-      {
-        title: 'Likey',
-      },
-      {
-        title: 'Turtle',
-      },
-      {
-        title: 'Missing U',
-      },
-      {
-        title: 'Wow',
-      },
-      {
-        title: 'FFW',
-      },
-      {
-        title: 'Ding Dong',
-      },
-      {
-        title: '24/7',
-      },
-      {
-        title: 'Look At Me',
-      },
-      {
-        title: 'Rollin',
-      },
-      {
-        title: 'Love Line',
-      },
-      {
-        title: "Don't Give Up",
-      },
-      {
-        title: 'You In My Heart',
-      },
-      {
-        title: 'Jaljayo Good Night',
-      },
-    ],
-  },
-  {
-    id: 'candy-pop',
-    name: 'Candy Pop',
-    image: candyPop,
-    unit: [Unit.twice],
-    types: ['single'],
-    market: 'japanese',
-    releaseDate: new Date('2018-02-07'),
-    songs: [
-      {
-        title: 'Candy Pop',
-        link: 'https://youtu.be/wQ_POfToaVY?si=LiBbD9ZlpktmQlTM',
-      },
-      {
-        title: 'Brand New Girl',
-        link: 'https://youtu.be/r1CMjQ0QJ1E?si=PhTGMPkXcelnh7AK',
-      },
-      {
-        title: 'Candy Pop (Instrumental)',
-      },
-      {
-        title: 'Brand New Girl (Instrumental)',
-      },
-    ],
-  },
-  {
-    id: 'what-is-love',
-    name: 'What is Love?',
-    image: whatIsLove,
-    unit: [Unit.twice],
-    releaseDate: new Date('2018-04-09'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'What is Love?',
-        link: 'https://youtu.be/i0p1bmr0EmE?si=fwBY53P6-0kWkzNA',
-      },
-      {
-        title: 'Sweet Talker',
-      },
-      {
-        title: 'Ho!',
-      },
-      {
-        title: 'Dejavu',
-      },
-      {
-        title: 'Say Yes',
-      },
-      {
-        title: 'Stuck',
-      },
-    ],
-  },
-  {
-    id: 'wake-me-up',
-    name: 'Wake Me Up',
-    image: wakeMeUp,
-    unit: [Unit.twice],
-    releaseDate: new Date('2018-05-16'),
-    types: ['single'],
-    market: 'japanese',
-    songs: [
-      {
-        title: 'Wake Me Up',
-        link: 'https://youtu.be/DdLYSziSXII?si=d9dunDZ07Hw6C17C',
-      },
-      {
-        title: 'Pink Lemonade',
-      },
-      {
-        title: 'Wake Me Up (Instrumental)',
-      },
-      {
-        title: 'Pink Lemonade (Instrumental)',
-      },
-    ],
-  },
-  {
-    id: 'i-want-you-back',
-    name: 'I Want You Back',
-    image: iWantYouBack,
-    unit: [Unit.twice],
-    types: ['single'],
-    market: 'japanese',
-    releaseDate: new Date('2018-06-15'),
-    songs: [
-      {
-        title: 'I Want You Back',
-        link: 'https://youtu.be/4OjLyOyMcNg?si=GEAoRk6t9ZzovHhO',
-      },
-    ],
-  },
-  {
-    id: 'summer-nights',
-    name: 'Summer Nights',
-    image: summerNights,
-    unit: [Unit.twice],
-    releaseDate: new Date('2018-07-09'),
-    types: ['mini-album', 'repackage'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'Dance The Night Away',
-        link: 'https://youtu.be/Fm5iP0S1z9w?si=5a0mXhJwlx4Up64O',
-      },
-      {
-        title: 'Chillax',
-      },
-      {
-        title: 'Shot Thru the Heart',
-      },
-      {
-        title: 'What is Love?',
-      },
-      {
-        title: 'Sweet Talker',
-      },
-      {
-        title: 'Ho!',
-      },
-      {
-        title: 'Dejavu',
-      },
-      {
-        title: 'Say Yes',
-      },
-      {
-        title: 'Stuck',
-      },
-    ],
-  },
-  {
-    id: 'bdz',
-    name: 'BDZ',
-    image: bdz,
-    unit: [Unit.twice],
-    releaseDate: new Date('2018-09-12'),
-    types: ['full-album'],
-    market: 'japanese',
-    songs: [
-      {
-        title: 'BDZ',
-        link: 'https://youtu.be/CMNahhgR_ss?si=TxQ3r6j0cU8g3AqS',
-      },
-      {
-        title: 'One More Time',
-        link: 'https://youtu.be/HuoOEry-Yc4?si=bYc5JxHj5w2Dbe3Z',
-      },
-      {
-        title: 'Candy Pop',
-        link: 'https://youtu.be/wQ_POfToaVY?si=J9TtO0Urx54vmyUg',
-      },
-      {
-        title: 'L.O.V.E.',
-      },
-      {
-        title: 'Wishing',
-      },
-      {
-        title: 'Say It Again',
-      },
-      {
-        title: 'Wake Me Up',
-      },
-      {
-        title: 'Brand New Girl',
-      },
-      {
-        title: 'Be as ONE',
-        link: 'https://youtu.be/fnb1q3l4yLM?si=bRHRgI_B9tW2H7Rb',
-      },
-      {
-        title: 'I Want You Back',
-      },
-    ],
-  },
-  {
-    id: 'stay-by-my-side',
-    name: 'Stay By My Side',
-    image: stayByMySide,
-    unit: [Unit.twice],
-    releaseDate: new Date('2018-10-22'),
-    types: ['single'],
-    market: 'japanese',
-    songs: [
-      {
-        title: 'Stay By My Side',
-        link: 'https://youtu.be/diIoA6UquMw?si=09HzNK_g0YNO8ZZu',
-      },
-    ],
-  },
-  {
-    id: 'yes-or-yes',
-    name: 'YES or YES',
-    image: yesOrYes,
-    unit: [Unit.twice],
-    releaseDate: new Date('2018-11-05'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'YES or YES',
-        link: 'https://youtu.be/mAKsZ26SabQ?si=kFVSi8cB8mUqQm2i',
-      },
-      { title: 'Say You Love Me' },
-      { title: 'Lalala' },
-      { title: 'Young & Wild' },
-      { title: 'Sunset' },
-      { title: 'After Moon' },
-      { title: 'BDZ (Korean Ver.)' },
-    ],
-  },
-  {
-    id: 'the-year-of-yes',
-    name: 'The Year of “YES”',
-    image: theYearOfYes,
-    unit: [Unit.twice],
-    releaseDate: new Date('2018-12-12'),
-    types: ['mini-album', 'repackage'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'The Best Thing I Ever Did',
-        link: 'https://youtu.be/xmJ3Lu7Vs9w?si=3U6uZr24UEfj7uBd',
-      },
-      { title: 'Be as ONE (Korean Ver.)' },
-      { title: 'YES or YES' },
-      { title: 'Say You Love Me' },
-      { title: 'Lalala' },
-      { title: 'Young & Wild' },
-      { title: 'Sunset' },
-      { title: 'After Moon' },
-      { title: 'BDZ (Korean Ver.)' },
-    ],
-  },
-  {
-    id: 'bdz-repackage',
-    name: 'BDZ -Repackage-',
-    image: bdz,
-    unit: [Unit.twice],
-    releaseDate: new Date('2018-12-26'),
-    types: ['full-album', 'repackage'],
-    market: 'japanese',
-    songs: [
-      {
-        title: 'Stay By My Side',
-        link: 'https://youtu.be/96K5RxgTfW4?si=0wKUDSoa2yYuXKA2',
-      },
-      { title: 'BDZ' },
-      { title: 'One More Time' },
-      { title: 'Candy Pop' },
-      { title: 'L.O.V.E.' },
-      { title: 'Wishing' },
-      { title: 'Say It Again' },
-      { title: 'Wake Me Up' },
-      { title: 'Brand New Girl' },
-      { title: 'Be as ONE' },
-      { title: 'I Want You Back' },
-    ],
-  },
-  {
-    id: 'hashtag-twice2',
-    name: '#TWICE2',
-    image: hashtagTwice2,
-    unit: [Unit.twice],
-    types: ['mini-album', 'repackage'],
-    market: 'japanese',
-    releaseDate: new Date('2019-03-06'),
-    songs: [
-      {
-        title: 'Likey (Japanese Ver.)',
-        link: 'https://youtu.be/N7MKlhS2ysU?si=kqxtKdOwtkIfjkAi',
-      },
-      { title: 'Heart Shaker (Japanese Ver.)' },
-      {
-        title: 'What Is Love? (Japanese Ver.)',
-        link: 'https://youtu.be/3zQXMPbK5jU?si=vGFCzwdzrNfBqpPk',
-      },
-      { title: 'Dance the Night Away (Japanese Ver.)' },
-      { title: 'Yes or Yes (Japanese Ver.)' },
-      { title: 'Likey' },
-      { title: 'Heart Shaker' },
-      { title: 'What Is Love?' },
-      { title: 'Yes or Yes' },
-    ],
-  },
-  {
-    id: 'fancy-you',
-    name: 'Fancy You',
-    image: fancyYou,
-    unit: [Unit.twice],
-    releaseDate: new Date('2019-04-22'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'Fancy',
-        link: 'https://youtu.be/kOHB85vDuow?si=VCPsEmIo1zV0lM7s',
-      },
-      { title: 'Stuck in My Head' },
-      { title: 'Girls Like Us' },
-      { title: 'Hot' },
-      { title: 'Turn it Up' },
-      { title: 'Strawberry' },
-    ],
-  },
-  {
-    id: 'happy-happy',
-    name: 'Happy Happy',
-    image: happyHappy,
-    unit: [Unit.twice],
-    releaseDate: new Date('2019-07-17'),
-    types: ['single'],
-    market: 'japanese',
-    songs: [
-      {
-        title: 'Happy Happy',
-        link: 'https://youtu.be/3n9rDwpa6QA?si=3pL6mGkeu4QIdg6E',
-      },
-      { title: 'The Best Thing I Ever Did (Japanese Ver.)' },
-      { title: 'Happy Happy (Collapsedone Remix)' },
-      { title: 'Happy Happy (Instrumental)' },
-    ],
-  },
-  {
-    id: 'breakthrough',
-    name: 'Breakthrough',
-    image: breakthrough,
-    unit: [Unit.twice],
-    releaseDate: new Date('2019-07-24'),
-    types: ['single'],
-    market: 'japanese',
-    songs: [
-      {
-        title: 'Breakthrough',
-        link: 'https://youtu.be/1j6YA03hm4k?si=31jBdN9TifhhyY6K',
-      },
-      { title: 'Fancy (Japanese Ver.)' },
-      { title: 'Breakthrough (Taalthechoi Remix)' },
-      { title: 'Breakthrough (Instrumental)' },
-    ],
-  },
-  {
-    id: 'feel-special',
-    name: 'Feel Special',
-    image: feelSpecial,
-    unit: [Unit.twice],
-    releaseDate: new Date('2019-09-23'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'Feel Special',
-        link: 'https://youtu.be/3ymwOvzhwHs?si=F3yObIMy2T4BPrdM',
-      },
-      { title: 'Rainbow' },
-      { title: 'Get Loud' },
-      { title: 'Trick It' },
-      { title: 'Love Foolish' },
-      { title: '21:29' },
-      { title: 'Breakthrough (Korean Ver.)' },
-    ],
-  },
-  {
-    id: 'and-twice',
-    name: '&TWICE',
-    image: andTwice,
-    unit: [Unit.twice],
-    releaseDate: new Date('2019-11-20'),
-    types: ['full-album'],
-    market: 'japanese',
-    songs: [
-      {
-        title: 'Fake & True',
-        link: 'https://youtu.be/zQELp93xxfo?si=b3i0dd8I28fjs4Pt',
-      },
-      { title: 'Stronger' },
-      { title: 'Breakthrough' },
-      { title: 'Changing!' },
-      { title: 'Happy Happy' },
-      { title: 'What You Waiting For' },
-      { title: 'Be OK' },
-      { title: 'POLISH' },
-      { title: 'How u doin' },
-      { title: 'The Reason Why' },
-    ],
-  },
-  {
-    id: 'and-twice-repackage',
-    name: '&TWICE -Repackage',
-    image: andTwice,
-    unit: [Unit.twice],
-    releaseDate: new Date('2020-02-05'),
-    types: ['full-album', 'repackage'],
-    market: 'japanese',
-    songs: [
-      { title: 'Swing' },
-      { title: 'Fake & True' },
-      { title: 'Stronger' },
-      { title: 'Breakthrough' },
-      { title: 'Changing!' },
-      { title: 'Happy Happy' },
-      { title: 'What You Waiting For' },
-      { title: 'Be OK' },
-      { title: 'POLISH' },
-      { title: 'How u doin' },
-      { title: 'The Reason Why' },
-    ],
-  },
-  {
-    id: 'more-and-more',
-    name: 'More & More',
-    image: moreAndMore,
-    unit: [Unit.twice],
-    releaseDate: new Date('2020-06-01'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'More & More',
-        link: 'https://youtu.be/mH0_XpSHkZo?si=nl1WIVJKwKn4Kf_e',
-      },
-      { title: 'Oxygen' },
-      { title: 'Firework' },
-      { title: 'Make Me Go' },
-      { title: 'Shadow' },
-      { title: "Don't Call Me Again" },
-      { title: 'Sweet Summer Day' },
-    ],
-  },
-  {
-    id: 'fanfare',
-    name: 'Fanfare',
-    image: fanfare,
-    unit: [Unit.twice],
-    releaseDate: new Date('2020-07-08'),
-    types: ['single'],
-    market: 'japanese',
-    songs: [
-      { title: 'Fanfare' },
-      { title: 'More & More (Japanese Ver.)' },
-      { title: 'Fanfare (Lee Hae Sol Remix)' },
-      { title: 'Fanfare (Instrumental)' },
-    ],
-  },
-  {
-    id: 'more-and-more-english',
-    name: 'More & More (English Ver.)',
-    image: moreAndMoreEnglish,
-    unit: [Unit.twice],
-    types: ['single'],
-    market: 'english',
-    releaseDate: new Date('2020-08-21'),
-    songs: [{ title: 'More & More (English Ver.)' }],
-  },
-  {
-    id: 'hashtag-twice3',
-    name: '#TWICE3',
-    image: hashtagTwice3,
-    unit: [Unit.twice],
-    types: ['mini-album'],
-    market: 'japanese',
-    releaseDate: new Date('2020-09-16'),
-    songs: [
-      { title: 'The Best Thing I Ever Did (Japanese Ver.)' },
-      { title: 'Fancy (Japanese Ver.)' },
-      { title: 'Feel Special (Japanese Ver.)' },
-      { title: 'More & More (Japanese Ver.)' },
-      { title: 'Stuck In My Head (Japanese Ver.)' },
-      { title: '21:29 (Japanese Ver.)' },
-      { title: 'The Best Thing I Ever Did' },
-      { title: 'Fancy' },
-      { title: 'Feel Special' },
-      { title: 'More & More' },
-      { title: 'Stuck In My Head' },
-      { title: '21:29' },
-    ],
-  },
-  {
-    id: 'eyes-wide-open',
-    name: 'Eyes Wide Open',
-    image: eyesWideOpen,
-    unit: [Unit.twice],
-    types: ['full-album'],
-    market: 'korean',
-    releaseDate: new Date('2020-10-26'),
-    songs: [
-      {
-        title: "I Can't Stop Me",
-        link: 'https://youtu.be/CM4CkVFmTds?si=E4MnHiQ2VRAxt0c8',
-      },
-      { title: 'Hell in Heaven' },
-      { title: 'Up No More' },
-      { title: 'Do What We Like' },
-      { title: 'Bring It Back' },
-      { title: 'Believer' },
-      { title: 'Queen' },
-      { title: 'Go Hard' },
-      { title: 'Shot Clock' },
-      { title: 'Handle It' },
-      { title: 'Depend On You' },
-      { title: 'Say Something' },
-      { title: 'Behind The Mask' },
-    ],
-  },
-  {
-    id: 'better',
-    name: 'BETTER',
-    image: better,
-    unit: [Unit.twice],
-    types: ['single'],
-    market: 'japanese',
-    releaseDate: new Date('2020-11-18'),
-    songs: [
-      { title: 'BETTER' },
-      { title: 'Scorpion' },
-      { title: 'BETTER (Instrumental)' },
-      { title: 'Scorpion (Instrumental)' },
-    ],
-  },
-  {
-    id: 'i-cant-stop-me-english',
-    name: "I Can't Stop Me (English Ver.)",
-    image: iCantStopMeEnglish,
-    unit: [Unit.twice],
-    types: ['single', 'digital'],
-    market: 'english',
-    releaseDate: new Date('2020-11-30'),
-    songs: [{ title: "I Can't Stop Me (English Ver.)" }],
-  },
-  {
-    id: 'cry-for-me',
-    name: 'Cry For Me',
-    image: cryForMe,
-    unit: [Unit.twice],
-    releaseDate: new Date('2020-12-18'),
-    types: ['single'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'Cry For Me',
-        link: 'https://youtu.be/bkQw-F1QTq4?si=2E5gkC0oq5LOnHyK',
-      },
-    ],
-  },
-  {
-    id: 'kura-kura',
-    name: 'Kura Kura',
-    image: kuraKura,
-    unit: [Unit.twice],
-    releaseDate: new Date('2021-05-12'),
-    types: ['single'],
-    market: 'japanese',
-    songs: [
-      {
-        title: 'Kura Kura',
-        link: 'https://youtu.be/BSS8Y-0hOlY?si=CbwKOkJK3_EFr1ho',
-      },
-      { title: 'Strawberry Moon' },
-      { title: 'Kura Kura (Instrumental)' },
-      { title: 'Strawberry Moon (Instrumental)' },
-    ],
-  },
-  {
-    id: 'taste-of-love',
-    name: 'Taste Of Love',
-    image: tasteOfLove,
-    unit: [Unit.twice],
-    releaseDate: new Date('2021-06-11'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'Alcohol-Free',
-        link: 'https://youtu.be/X-uJtV8ScYk?si=PCV4H2zC3N2uq5Dw',
-      },
-      { title: 'First Time' },
-      { title: 'Scandal' },
-      { title: 'Conversation' },
-      { title: 'Baby Blue Love' },
-      { title: 'SOS' },
-      { title: 'Cry For Me (English ver.)' },
-    ],
-  },
-  {
-    id: 'hospital-playlist-season2',
-    name: 'HOSPITAL PLAYLIST Season2, Pt. 4',
-    image: hospitalPlaylistSeason2,
-    unit: [Unit.twice],
-    releaseDate: new Date('2021-07-09'),
-    types: ['single', 'ost'],
-    market: 'korean',
-    songs: [
-      { title: 'I love you more than anyone' },
-      { title: 'I love you more than anyone - Instrumental' },
-    ],
-  },
-  {
-    id: 'perfect-world',
-    name: 'Perfect World',
-    image: perfectWorld,
-    unit: [Unit.twice],
-    releaseDate: new Date('2021-07-28'),
-    types: ['full-album'],
-    market: 'japanese',
-    songs: [
-      {
-        title: 'Perfect World',
-        link: 'https://youtu.be/fmOEKOjyDxU?si=IOp-WqxWKoVhZ29E',
-      },
-      { title: 'BETTER' },
-      { title: 'Good at Love' },
-      { title: 'Fanfare' },
-      { title: 'Kura Kura' },
-      { title: 'Four-leaf Clover' },
-      { title: 'In the summer' },
-      { title: 'PIECES OF LOVE' },
-      { title: 'Thank you, Family' },
-      { title: 'PROMISE' },
-    ],
-  },
-  {
-    id: 'the-feels',
-    name: 'The Feels',
-    image: theFeels,
-    unit: [Unit.twice],
-    releaseDate: new Date('2021-10-01'),
-    types: ['single'],
-    market: 'english',
-    songs: [
-      {
-        title: 'The Feels',
-        link: 'https://youtu.be/f5_wn8mexmM?si=m22mkuFgJPO07PsZ',
-      },
-      { title: 'The Feels (Stereotypes Remix)' },
-      { title: 'The Feels (YVES V Remix)' },
-      { title: 'The Feels (Instrumental)' },
-      { title: 'The Feels (The Stereotypes Remix [Instrumental])' },
-      { title: 'The Feels (YVES V Remix [Instrumental])' },
-    ],
-  },
-  {
-    id: 'formula-of-love',
-    name: 'Formula of Love: O+T=<3',
-    image: formulaOfLove,
-    unit: [Unit.twice],
-    releaseDate: new Date('2021-11-12'),
-    types: ['full-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'Scientist',
-        link: 'https://youtu.be/vPwaXytZcgI?si=zOWqgL_FdJX_37jZ',
-      },
-      { title: 'Moonlight' },
-      { title: 'Icon' },
-      { title: 'Cruel' },
-      { title: 'Real You' },
-      { title: 'F.I.L.A (Fall In Love Again)' },
-      { title: 'Last Waltz' },
-      { title: 'Espresso' },
-      { title: 'Rewind' },
-      { title: 'Cactus' },
-      { title: 'Push & Pull' },
-      { title: 'Hello' },
-      { title: '1, 3, 2' },
-      { title: 'Candy' },
-      { title: 'The Feels' },
-      { title: 'Scientist (R3HAB Remix)' },
-      { title: 'The Feels (Korean Ver)' },
-    ],
-  },
-  {
-    id: 'doughnut',
-    name: 'Doughnut',
-    image: doughnut,
-    unit: [Unit.twice],
-    releaseDate: new Date('2021-12-15'),
-    types: ['single'],
-    market: 'japanese',
-    songs: [
-      {
-        title: 'Doughnut',
-        link: 'https://youtu.be/VcOSUOpACq0?si=1ijWUF1QRp51rnWu',
-      },
-      { title: 'Wonderful Day' },
-      { title: 'Doughnut (Instrumental)' },
-      { title: 'Wonderful Day (Instrumental)' },
-    ],
-  },
-  {
-    id: 'hashtag-twice4',
-    name: '#TWICE4 (Japanese ver.)',
-    image: hashtagTwice4,
-    unit: [Unit.twice],
-    releaseDate: new Date('2022-03-16'),
-    types: ['mini-album'],
-    market: 'japanese',
-    songs: [
-      { title: "I Can't Stop Me - Japanese ver." },
-      { title: 'Cry For Me - Japanese ver.' },
-      { title: 'Alcohol-Free - Japanese ver.' },
-      { title: 'Scientist - Japanese ver.' },
-    ],
-  },
-  {
-    id: 'just-be-yourself',
-    name: 'Just Be Yourself',
-    image: justBeYourself,
-    unit: [Unit.twice],
-    releaseDate: new Date('2022-03-22'),
-    types: ['single'],
-    market: 'japanese',
-    songs: [{ title: 'Just Be Yourself' }],
-  },
-  {
-    id: 'celebrate',
-    name: 'Celebrate',
-    image: celebrate,
-    unit: [Unit.twice],
-    releaseDate: new Date('2022-07-27'),
-    types: ['full-album'],
-    market: 'japanese',
-    songs: [
-      {
-        title: 'Celebrate',
-        link: 'https://youtu.be/fMIn43MiwG8?si=UwqnJmtpk6r1kX2f',
-      },
-      { title: 'Voices of Delight' },
-      { title: 'TICK TOCK' },
-      { title: 'Flow like waves' },
-      { title: "That's all I'm saying" },
-      { title: 'Bitter Sweet' },
-      { title: 'Sandcastle' },
-      { title: 'Just Be Yourself' },
-      { title: 'Doughnut' },
-    ],
-  },
-  {
-    id: 'between1and2',
-    name: 'Between 1&2',
-    image: between1And2,
-    unit: [Unit.twice],
-    releaseDate: new Date('2022-08-26'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'Talk That Talk',
-        link: 'https://youtu.be/k6jqx9kZgPM?si=LLCGQkDlhOsuAcGv',
-      },
-      { title: 'Queen Of Hearts' },
-      { title: 'Basics' },
-      { title: 'Trouble' },
-      { title: 'Brave' },
-      { title: 'Gone' },
-      { title: 'When We Were Kids' },
-    ],
-  },
-  {
-    id: 'moonlight-sunrise',
-    name: 'Moonlight Sunrise',
-    image: moonlightSunrise,
-    unit: [Unit.twice],
-    releaseDate: new Date('2023-01-20'),
-    types: ['single', 'digital'],
-    market: 'english',
-    songs: [
-      {
-        title: 'Moonlight Sunrise',
-        link: 'https://youtu.be/cKlEE_EYuNM?si=Bjib_kTGEoq2r_5Z',
-      },
-    ],
-  },
-  {
-    id: 'ready-to-be',
-    name: 'READY TO BE',
-    image: readyToBe,
-    unit: [Unit.twice],
-    releaseDate: new Date('2023-03-10'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'SET ME FREE',
-        link: 'https://youtu.be/w4cTYnOPdNk?si=E-O2l-zXYMUWXA2q',
-      },
-      { title: 'MOONLIGHT SUNRISE' },
-      { title: 'GOT THE THRILLS' },
-      { title: 'BLAME IT ON ME' },
-      { title: 'WALLFLOWER' },
-      { title: 'CRAZY STUPID LOVE' },
-      { title: 'SET ME FREE (Eng. Ver.)' },
-    ],
-  },
-  {
-    id: 'hare-hare',
-    name: 'Hare Hare',
-    image: hareHare,
-    unit: [Unit.twice],
-    releaseDate: new Date('2023-05-31'),
-    types: ['mini-album'],
-    market: 'japanese',
-    songs: [
-      {
-        title: 'Hare Hare',
-        link: 'https://youtu.be/-uqWaGzQyxA?si=tZKruWEMeWQ1F5fL',
-      },
-      { title: 'Catch a Wave' },
-      { title: 'Hare Hare (Instrumental)' },
-      { title: 'Catch a Wave (Instrumental)' },
-    ],
-  },
-  {
-    id: 'heart-shaker-galactika-holiday-remix',
-    name: 'Heart Shaker (GALACTIKA * Holiday Remix)',
-    image: heartShakerGalactikaHolidayRemix,
-    unit: [Unit.twice],
-    releaseDate: new Date('2023-11-07'),
-    types: ['single'],
-    market: 'korean',
-    songs: [{ title: 'Heart Shaker (GALACTIKA * Holiday Remix)' }],
-  },
-  {
-    id: 'the-remixes',
-    name: 'THE REMIXES',
-    image: theRemixes,
-    unit: [Unit.twice],
-    releaseDate: new Date('2023-11-22'),
-    types: ['single'],
-    market: 'korean',
-    songs: [
-      { title: 'MOONLIGHT SUNRISE (Jonas Blue Remix)' },
-      { title: 'The Feels (Ian Asher Remix)' },
-      { title: 'SET ME FREE (Carneyval Remix)' },
-      { title: 'Alcohol-Free (English ver.)' },
-      { title: 'CRY FOR ME (English ver.)' },
-      { title: "I CAN'T STOP ME (ft.BOYS LIKE GIRLS)" },
-      { title: 'MORE & MORE (LEE HAE SOL Sped Up Remix)' },
-    ],
-  },
-  {
-    id: 'dance-again',
-    name: 'Dance Again',
-    image: danceAgain,
-    unit: [Unit.twice],
-    releaseDate: new Date('2023-12-12'),
-    types: ['single'],
-    market: 'japanese',
-    songs: [
-      {
-        title: 'Dance Again',
-        link: 'https://youtu.be/OzmnjIca8LU?si=jIFDjlayXWi2doij',
-      },
-    ],
-  },
-  {
-    id: 'i-got-you',
-    name: 'I GOT YOU',
-    image: iGotYou,
-    unit: [Unit.twice],
-    releaseDate: new Date('2024-02-02'),
-    types: ['single'],
-    market: 'korean',
-    songs: [{ title: 'I GOT YOU' }],
-  },
-  {
-    id: 'i-got-you-voyage-ver',
-    name: 'I GOT YOU  (Voyage ver.)',
-    image: iGotYouVoyageVer,
-    unit: [Unit.twice],
-    releaseDate: new Date('2024-02-05'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      { title: 'I GOT YOU (Voyage ver.)' },
-      { title: 'I GOT YOU (Feat. Lauv)' },
-      { title: 'I GOT YOU (Original)' },
-      { title: 'I GOT YOU (Lo-fi ver.)' },
-      { title: 'I GOT YOU (Hyper ver.)' },
-      { title: 'I GOT YOU (Garage ver.)' },
-      { title: 'I GOT YOU (Sped Up ver.)' },
-      { title: 'I GOT YOU (Feat. Lauv) [Instrumental]' },
-      { title: 'I GOT YOU (Original) [Instrumental]' },
-      { title: 'I GOT YOU (Lo-fi ver.) [Instrumental]' },
-      { title: 'I GOT YOU (Hyper ver.) [Instrumental]' },
-      { title: 'I GOT YOU (Garage ver.) [Instrumental]' },
-      { title: 'I GOT YOU (Sped Up ver.) [Instrumental]' },
-    ],
-  },
-  {
-    id: 'with-you-th',
-    name: 'With YOU-th',
-    image: withYouTh,
-    unit: [Unit.twice],
-    releaseDate: new Date('2024-02-23'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'I GOT YOU',
-        link: 'https://youtu.be/haf67eKF0uo?si=2IqcPS6dAeaf01a3',
-      },
-      {
-        title: 'ONE SPARK',
-        link: 'https://youtu.be/jCzez_q8si0?si=yPGAODsp98RCPX57',
-      },
-      { title: 'RUSH' },
-      { title: 'NEW NEW' },
-      { title: 'BLOOM' },
-      { title: 'YOU GET ME' },
-    ],
-  },
-  {
-    id: 'one-spark',
-    name: 'ONE SPARK',
-    image: oneSpark,
-    unit: [Unit.twice],
-    releaseDate: new Date('2024-02-26'),
-    types: ['single'],
-    market: 'korean',
-    songs: [
-      { title: 'ONE SPARK (English ver.)' },
-      { title: 'ONE SPARK (House ver.)' },
-      { title: 'ONE SPARK (R&B ver.)' },
-      { title: 'ONE SPARK (Tropical ver.)' },
-      { title: 'ONE SPARK (Rave ver.)' },
-      { title: 'ONE SPARK (Instrumental)' },
-    ],
-  },
-  {
-    id: 'dive',
-    name: 'DIVE',
-    image: dive,
-    unit: [Unit.twice],
-    releaseDate: new Date('2024-07-17'),
-    types: ['full-album'],
-    market: 'japanese',
-    songs: [
-      { title: 'Beyond the Horizon' },
-      { title: 'DIVE' },
-      { title: 'Ocean Deep' },
-      { title: 'LOVE WARNING' },
-      { title: 'Here I Am' },
-      { title: 'Inside of me' },
-      { title: 'Peach Soda' },
-      { title: 'Echoes of Heart' },
-      { title: 'Dance Again' },
-      { title: 'Hare Hare' },
-    ],
-  },
-  {
-    id: 'mamushi-remix-feat-twice',
-    name: 'Mamushi (Remix) [feat. TWICE]',
-    image: mamushiRemixFeatTwice,
-    unit: [Unit.twice],
-    releaseDate: new Date('2024-10-24'),
-    types: ['single'],
-    market: 'english',
-    songs: [{ title: 'Mamushi (Remix) [feat. TWICE]' }],
-  },
-  {
-    id: 'strategy',
-    name: 'STRATEGY',
-    image: strategy,
-    unit: [Unit.twice],
-    releaseDate: new Date('2024-12-06'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'Strategy (feat. Megan Thee Stallion)',
-        link: 'https://youtu.be/Sz_wWzgh-vQ?si=gf_2ADfNKPvFAHts',
-      },
-      { title: 'Kiss My Troubles Away' },
-      { title: 'Like It Like It' },
-      { title: 'Sweetest Obsession' },
-      { title: 'Keeper' },
-      { title: 'Magical' },
-      { title: 'Strategy' },
-    ],
-  },
-  {
-    id: 'the-wish',
-    name: 'The wish',
-    image: theWish,
-    unit: [Unit.twice],
-    releaseDate: new Date('2024-12-16'),
-    types: ['single'],
-    market: 'japanese',
-    songs: [{ title: 'The wish' }],
-  },
-  {
-    id: 'strategy2',
-    name: 'Strategy 2.0',
-    image: strategy2,
-    unit: [Unit.twice],
-    releaseDate: new Date('2024-12-18'),
-    types: ['single'],
-    market: 'english',
-    songs: [
-      { title: 'Strategy (version 1.0)' },
-      { title: 'Strategy (Slom Remix)' },
-      { title: 'Strategy (House)' },
-      { title: 'Strategy (Moombahton)' },
-      { title: 'Strategy (Inst.)' },
-    ],
-  },
-  {
-    id: 'we-pray-twice-version',
-    name: 'WE PRAY (TWICE Version)',
-    image: wePrayTwiceVersion,
-    unit: [Unit.twice],
-    releaseDate: new Date('2025-04-18'),
-    types: ['single'],
-    market: 'english',
-    songs: [
-      {
-        title: 'WE PRAY (TWICE Version) (feat. Little Simz, Burna Boy, Elyanna)',
-      },
-    ],
-  },
-  {
-    id: 'hashtag-twice5',
-    name: '#TWICE5',
-    image: hashtagTwice5,
-    unit: [Unit.twice],
-    releaseDate: new Date('2025-05-14'),
-    types: ['mini-album'],
-    market: 'japanese',
-    songs: [
-      { title: 'Talk that Talk (Japanese ver.)' },
-      { title: 'SET ME FREE (Japanese ver.)' },
-      { title: 'ONE SPARK (Japanese ver.)' },
-      { title: 'Talk that Talk' },
-      { title: 'SET ME FREE' },
-      { title: 'I GOT YOU' },
-      { title: 'ONE SPARK' },
-      { title: 'Strategy' },
-    ],
-  },
-  {
-    id: 'kpop-demon-hunters',
-    name: 'KPop Demon Hunters',
-    image: kpopDemonHunters,
-    unit: [Unit.jeongyeon, Unit.jihyo, Unit.chaeyoung],
-    releaseDate: new Date('2025-06-20'),
-    types: ['single', 'ost'],
-    market: 'korean',
-    songs: [{ title: 'TAKEDOWN (JEONGYEON, JIHYO, CHAEYOUNG)' }, { title: 'Strategy' }],
-  },
-  {
-    id: 'this-is-for',
-    name: 'THIS IS FOR',
-    image: thisIsFor,
-    unit: [Unit.twice],
-    releaseDate: new Date('2025-07-11'),
-    types: ['full-album'],
-    market: 'korean',
-    songs: [
-      { title: 'FOUR' },
-      {
-        title: 'THIS IS FOR',
-        link: 'https://youtu.be/eHHQaoEW30Q?si=apjBn-LzQUcQDOK8',
-      },
-      { title: 'OPTIONS' },
-      { title: 'MARS' },
-      { title: 'RIGHT HAND GIRL' },
-      { title: 'PEACH GELATO' },
-      { title: 'HI HELLO' },
-      { title: 'BATTITUDE (NAYEON, JEONGYEON, MOMO, MINA)' },
-      { title: 'DAT AAH DAT OOH (SANA, JIHYO, DAHYUN, CHAEYOUNG, TZUYU)' },
-      { title: 'LET LOVE GO (JEONGYEON, MOMO, SANA, TZUYU)' },
-      { title: 'G.O.A.T (MINA, DAHYUN, CHAEYOUNG)' },
-      { title: 'TALK (NAYEON, JIHYO)' },
-      { title: 'SEESAW' },
-    ],
-  },
-  {
-    id: 'this-is-for-deluxe',
-    name: 'THIS IS FOR DELUXE',
-    image: thisIsForDeluxe,
-    unit: [Unit.twice],
-    releaseDate: new Date('2025-07-14'),
-    types: ['full-album'],
-    market: 'korean',
-    songs: [
-      { title: 'FOUR' },
-      { title: 'THIS IS FOR' },
-      { title: 'OPTIONS' },
-      { title: 'MARS' },
-      { title: 'RIGHT HAND GIRL' },
-      { title: 'PEACH GELATO' },
-      { title: 'HI HELLO' },
-      { title: 'BATTITUDE (NAYEON, JEONGYEON, MOMO, MINA)' },
-      { title: 'DAT AAH DAT OOH (SANA, JIHYO, DAHYUN, CHAEYOUNG, TZUYU)' },
-      { title: 'LET LOVE GO (JEONGYEON, MOMO, SANA, TZUYU)' },
-      { title: 'G.O.A.T (MINA, DAHYUN, CHAEYOUNG)' },
-      { title: 'TALK (NAYEON, JIHYO)' },
-      { title: 'SEESAW' },
-      { title: 'HEARTBREAK AVENUE' },
-      { title: 'THIS IS FOR (Extended)' },
-      { title: 'TAKEDOWN (JEONGYEON, JIHYO, CHAEYOUNG) (Kpop Demon Hunters)' },
-    ],
-  },
-  {
-    id: 'superstars-with-saweetie',
-    name: 'superstars (with Saweetie)',
-    image: superstarsWithSaweetie,
-    unit: [Unit.twice],
-    releaseDate: new Date('2025-08-01'),
-    types: ['single'],
-    market: 'english',
-    songs: [{ title: 'superstars (with Saweetie)' }],
-  },
-  {
-    id: 'enemy',
-    name: 'ENEMY',
-    image: enemy,
-    unit: [Unit.twice],
-    releaseDate: new Date('2025-08-27'),
-    types: ['full-album'],
-    market: 'japanese',
-    songs: [
-      { title: 'Up to you' },
-      { title: 'ENEMY' },
-      { title: 'FINE' },
-      { title: 'One day' },
-      { title: 'Blind in Love' },
-      { title: 'Love is more' },
-      { title: 'The wish' },
-      { title: 'Glow' },
-      { title: 'Like 1' },
-    ],
-  },
-  {
-    id: 'ten-the-story-goes-on',
-    name: 'TEN: The Story Goes On',
-    image: tenTheStoryGoesOn,
-    unit: [Unit.twice],
-    releaseDate: new Date('2025-10-10'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'ME+YOU',
-        link: 'https://youtu.be/zqorlX_5oHQ?si=GAYrtx85hFwfgH14',
-      },
-      { title: 'MEEEEEE (NAYEON)' },
-      { title: 'FIX A DRINK (JEONGYEON)' },
-      { title: 'MOVE LIKE THAT (MOMO)' },
-      { title: 'DECAFFEINATED (SANA)' },
-      { title: 'ATM (JIHYO)' },
-      { title: 'STONE COLD (MINA)' },
-      { title: 'CHESS (DAHYUN)' },
-      { title: 'IN MY ROOM (CHAEYOUNG)' },
-      { title: 'DIVE IN (TZUYU)' },
-    ],
-  },
-  {
-    id: 'daring-woman',
-    name: 'Daring Woman (Two Yoo Project Sugar Man OST)',
-    image: daringWoman,
-    unit: [Unit.nayeon, Unit.jihyo, Unit.chaeyoung, Unit.tzuyu],
-    releaseDate: new Date('2015-12-30'),
-    types: ['ost'],
-    market: 'korean',
-    songs: [{ title: '당돌한 여자 (Daring Woman)' }],
-  },
-  {
-    id: 'all-out',
-    name: 'ALL OUT',
-    image: allOut,
-    unit: [Unit.nayeon, Unit.sana, Unit.jihyo, Unit.chaeyoung],
-    releaseDate: new Date('2020-11-06'),
-    types: ['single'],
-    market: 'english',
-    songs: [{ title: "I'LL SHOW YOU" }],
-  },
-  {
-    id: 'im-nayeon',
-    name: 'IM NAYEON',
-    image: imNayeon,
-    unit: [Unit.nayeon],
-    releaseDate: new Date('2022-06-24'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'POP!',
-        link: 'https://youtu.be/f6YDKF0LVWw?si=R-FIdEgwSnqn1bae',
-      },
-      { title: 'NO PROBLEM (ft. Felix of Stray Kids)' },
-      { title: 'LOVE COUNTDOWN (ft. Wonstein)' },
-      { title: 'CANDYFLOSS' },
-      { title: 'ALL OR NOTHING' },
-      { title: 'HAPPY BIRTHDAY TO YOU' },
-      { title: '노을만 예쁘다 (SUNSET)' },
-    ],
-  },
-  {
-    id: 'na',
-    name: 'NA',
-    image: na,
-    unit: [Unit.nayeon],
-    releaseDate: new Date('2024-06-14'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'ABCD',
-        link: 'https://youtu.be/oUZttxRcPZw?si=J92rQanX91_ZV48R',
-      },
-      { title: 'Butterflies' },
-      { title: 'Heaven (ft. Sam Kim)' },
-      { title: 'Magic (ft. Julie of KISS OF LIFE)' },
-      { title: 'HalliGalli (ft. Lee Chanhyuk of AKMU)' },
-      { title: 'Something' },
-      { title: 'Count It' },
-    ],
-  },
-  {
-    id: 'abcd-extended-version',
-    name: 'ABCD (Extended Version)',
-    image: abcdExtendedVersion,
-    unit: [Unit.nayeon],
-    releaseDate: new Date('2024-06-24'),
-    types: ['single'],
-    market: 'korean',
-    songs: [
-      { title: 'ABCD' },
-      { title: 'ABCD (Extended Version)' },
-      { title: 'ABCD (Inst.)' },
-      { title: 'ABCD (Extended Version) (Inst.)' },
-    ],
-  },
-  {
-    id: 'beyond-end-credit-version',
-    name: 'Beyond (End Credit Version) [From “Moana 2”/Korean Soundtrack Version]',
-    image: beyondEndCreditVersion,
-    unit: [Unit.nayeon],
-    releaseDate: new Date('2024-11-20'),
-    types: ['single', 'ost'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'Beyond (End Credit Version) [From “Moana 2”/Korean Soundtrack Version]',
-        link: 'https://youtu.be/Kdubj6Rr-wE?si=wsfKObAKhRzRBSjv',
-      },
-    ],
-  },
-  {
-    id: 'kimagure-romantic',
-    name: 'Kimagure Romantic',
-    image: kimagureRomantic,
-    unit: [Unit.nayeon],
-    releaseDate: new Date('2025-05-30'),
-    types: ['single'],
-    market: 'japanese',
-    songs: [{ title: 'Kimagure Romantic' }],
-  },
-  {
-    id: 'era-with-nayeon',
-    name: 'Era with NAYEON (TWICE)',
-    image: eraWithNayeon,
-    unit: [Unit.nayeon],
-    releaseDate: new Date('2025-06-13'),
-    types: ['digital'],
-    market: 'english',
-    songs: [{ title: 'Era with NAYEON - (Remix)' }],
-  },
-  {
-    id: 'twenty-five-twenty-one',
-    name: '스물다섯 스물하나 OST (Twenty Five-Twenty-One)',
-    image: twentyFiveTwentyOne,
-    unit: [Unit.jihyo],
-    releaseDate: new Date('2022-03-06'),
-    types: ['ost'],
-    market: 'korean',
-    songs: [{ title: 'Stardust Love Song' }, { title: 'Stardust Love Song (Inst)' }],
-  },
-  {
-    id: 'zone',
-    name: 'ZONE (1st Mini Album)',
-    image: zone,
-    unit: [Unit.jihyo],
-    releaseDate: new Date('2023-08-18'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: "KILLIN' ME GOOD",
-        link: 'https://youtu.be/jyw_vrI4ySg?si=-06e3PHhFmWYms5B',
-      },
-      { title: "TALKIN' ABOUT IT (ft.24kGoldn)" },
-      { title: 'CLOSER' },
-      { title: 'WISHING ON YOU' },
-      { title: "DON'T WANNA GO BACK (Duet with Heize)" },
-      { title: 'ROOM' },
-      { title: 'NIGHTMARE' },
-    ],
-  },
-  {
-    id: 'killin-me-good-english',
-    name: "Killin' Me Good (English Ver.) (Single)",
-    image: killinMeGoodEnglishVer,
-    unit: [Unit.jihyo],
-    releaseDate: new Date('2023-09-15'),
-    types: ['single'],
-    market: 'english',
-    songs: [
-      { title: "Killin' Me Good (English Ver.)" },
-      { title: "TALKIN' ABOUT IT (ft.24kGoldn)" },
-    ],
-  },
-  {
-    id: 'marry-my-husband-ost-part2',
-    name: 'Marry My Husband OST Part.2',
-    image: marryMyHusbandOstPart2,
-    unit: [Unit.jihyo],
-    releaseDate: new Date('2024-01-06'),
-    types: ['ost'],
-    market: 'korean',
-    songs: [{ title: 'New Days' }, { title: 'New Days (Inst.)' }],
-  },
-  {
-    id: 'abou-tzu',
-    name: 'abouTZU (1st Mini Album)',
-    image: abouTzu,
-    unit: [Unit.tzuyu],
-    releaseDate: new Date('2024-09-06'),
-    types: ['mini-album'],
-    market: 'korean',
-    songs: [
-      {
-        title: 'Run Away',
-        link: 'https://youtu.be/IIrCrGAX03Y?si=P2xKLfNQPhAlSRe2',
-      },
-      { title: 'Heartbreak In Heaven (feat. PENIEL of BTOB)' },
-      { title: 'Lazy Baby (feat. pH-1)' },
-      { title: 'Losing Sleep' },
-      { title: 'One Love' },
-      { title: 'Fly' },
-    ],
-  },
-  {
-    id: 'blink',
-    name: 'BLINK (Corbyn Besson & TZUYU)',
-    image: blink,
-    unit: [Unit.tzuyu],
-    releaseDate: new Date('2025-09-26'),
-    types: ['digital', 'single'],
-    market: 'english',
-    songs: [
-      {
-        title: 'BLINK',
-        link: 'https://youtu.be/QcuV8h_I1y0?si=qNXn1U3lbIDuPjJP',
-      },
-    ],
-  },
-  {
-    id: 'lil-fantasy-vol1',
-    name: 'LIL FANTASY vol.1',
-    image: lilFantasyVol1,
-    unit: [Unit.chaeyoung],
-    releaseDate: new Date('2025-09-12'),
-    types: ['mini-album'],
-    market: 'english',
-    songs: [
-      {
-        title: 'Avocado (featuring Gliiico)',
-        link: 'https://youtu.be/NlHD4ndO-z0?si=wLlZYWKGTKyLvqqc',
-      },
-      { title: 'Band-Aid' },
-      {
-        title: 'Shoot (Firecracker)',
-        link: 'https://youtu.be/tl2luF-JRlU?si=SGzVaQfJcgwcV-H_',
-      },
-      { title: 'Girl' },
-      { title: 'Ribbons' },
-      { title: 'Downpour' },
-      { title: 'BF' },
-      { title: 'Shadow Puppet' },
-      { title: 'My Guitar' },
-    ],
-  },
-  {
-    id: 'bouquet',
-    name: 'Bouquet',
-    image: bouquet,
-    unit: [Unit.misamo],
-    releaseDate: new Date('2023-01-25'),
-    types: ['single', 'ost'],
-    market: 'japanese',
-    songs: [{ title: 'Bouquet' }],
-  },
-  {
-    id: 'masterpiece',
-    name: 'Masterpiece',
-    image: masterpiece,
-    unit: [Unit.misamo],
-    releaseDate: new Date('2023-07-26'),
-    types: ['mini-album'],
-    market: 'japanese',
-    songs: [
-      {
-        title: 'Do not touch',
-        link: 'https://youtu.be/woKq2sD8xho?si=YsP98bCgtSZ4K_y_',
-      },
-      { title: 'Behind The Curtain' },
-      { title: 'Marshmallow' },
-      { title: 'Funny Valentine' },
-      { title: "It's not easy for you" },
-      { title: 'Rewind you' },
-      { title: 'Bouquet' },
-    ],
-  },
-  {
-    id: 'haute-couture',
-    name: 'HAUTE COUTURE',
-    image: hauteCouture,
-    unit: [Unit.misamo],
-    releaseDate: new Date('2024-11-06'),
-    types: ['mini-album'],
-    market: 'japanese',
-    songs: [
-      { title: 'Identity' },
-      { title: 'RUNWAY' },
-      { title: 'Wah Wah Wah' },
-      { title: "Baby, I'm good" },
-      { title: 'Daydream' },
-      { title: 'NEW LOOK' },
-      { title: 'Jealousy' },
-      { title: 'Misty' },
-      { title: 'Money In My Pocket' },
-      { title: 'Mirage' },
-    ],
-  },
-  {
-    id: 'message',
-    name: 'Message',
-    image: message,
-    unit: [Unit.misamo],
-    releaseDate: new Date('2025-04-28'),
-    types: ['ost'],
-    market: 'japanese',
-    songs: [{ title: 'Message' }],
-  },
-  {
-    id: 'play',
-    name: 'PLAY',
-    image: play,
-    unit: [Unit.misamo],
-    releaseDate: new Date('2026-02-04'),
-    types: ['full-album'],
-    market: 'japanese',
-    songs: [
-      {
-        title: 'Confetti',
-        link: 'https://youtu.be/gOjG-70-D5o?si=6IdNucL19pb0XUsB',
-      },
-      { title: 'Hmm' },
-      { title: 'Not a Goodbye' },
-      { title: 'Deep Eden' },
-      { title: 'Red Diamond' },
-      { title: 'Catch My Eye' },
-      { title: 'Kitty (MOMO)' },
-      { title: 'Ma Cherry (SANA)' },
-      { title: 'Turning Tables (MINA)' },
-      { title: 'Identity' },
-      { title: 'Do not touch' },
-      { title: 'Message' },
-    ],
-  },
-]
+const ALBUM_IMAGES: Record<string, ImageMetadata> = {
+  'sotsugyou-cover': sotsugyouCover,
+  'like-a-star': likeaStar,
+  'i-fly': iFly,
+  'a-strange-day': aStrangeDay,
+  'popstar-remix': popStar,
+  'you-are-the-apple-of-my-eye': youAreTheAppleOfMyEye,
+  'the-story-begins': theStoryBegins,
+  'page-two': pageTwo,
+  'twicecoaster-lane-1': twicecoasterLane1,
+  'twicecoaster-lane-2': twicecoasterLane2,
+  'whats-this': whatsThis,
+  signal: signal,
+  'hashtag-twice': hashtagTwice,
+  'one-more-time': oneMoreTime,
+  twicetagram: twicetagram,
+  'merry-and-happy': merryAndHappy,
+  'candy-pop': candyPop,
+  'what-is-love': whatIsLove,
+  'wake-me-up': wakeMeUp,
+  'i-want-you-back': iWantYouBack,
+  'summer-nights': summerNights,
+  bdz: bdz,
+  'stay-by-my-side': stayByMySide,
+  'yes-or-yes': yesOrYes,
+  'the-year-of-yes': theYearOfYes,
+  'bdz-repackage': bdz,
+  'hashtag-twice2': hashtagTwice2,
+  'fancy-you': fancyYou,
+  'happy-happy': happyHappy,
+  breakthrough: breakthrough,
+  'feel-special': feelSpecial,
+  'and-twice': andTwice,
+  'and-twice-repackage': andTwice,
+  'more-and-more': moreAndMore,
+  fanfare: fanfare,
+  'more-and-more-english': moreAndMoreEnglish,
+  'hashtag-twice3': hashtagTwice3,
+  'eyes-wide-open': eyesWideOpen,
+  better: better,
+  'i-cant-stop-me-english': iCantStopMeEnglish,
+  'cry-for-me': cryForMe,
+  'kura-kura': kuraKura,
+  'taste-of-love': tasteOfLove,
+  'hospital-playlist-season2': hospitalPlaylistSeason2,
+  'perfect-world': perfectWorld,
+  'the-feels': theFeels,
+  'formula-of-love': formulaOfLove,
+  doughnut: doughnut,
+  'hashtag-twice4': hashtagTwice4,
+  'just-be-yourself': justBeYourself,
+  celebrate: celebrate,
+  between1and2: between1And2,
+  'moonlight-sunrise': moonlightSunrise,
+  'ready-to-be': readyToBe,
+  'hare-hare': hareHare,
+  'heart-shaker-galactika-holiday-remix': heartShakerGalactikaHolidayRemix,
+  'the-remixes': theRemixes,
+  'dance-again': danceAgain,
+  'i-got-you': iGotYou,
+  'i-got-you-voyage-ver': iGotYouVoyageVer,
+  'with-you-th': withYouTh,
+  'one-spark': oneSpark,
+  dive: dive,
+  'mamushi-remix-feat-twice': mamushiRemixFeatTwice,
+  strategy: strategy,
+  'the-wish': theWish,
+  strategy2: strategy2,
+  'we-pray-twice-version': wePrayTwiceVersion,
+  'hashtag-twice5': hashtagTwice5,
+  'kpop-demon-hunters': kpopDemonHunters,
+  'this-is-for': thisIsFor,
+  'this-is-for-deluxe': thisIsForDeluxe,
+  'superstars-with-saweetie': superstarsWithSaweetie,
+  enemy: enemy,
+  'ten-the-story-goes-on': tenTheStoryGoesOn,
+  'daring-woman': daringWoman,
+  'all-out': allOut,
+  'im-nayeon': imNayeon,
+  na: na,
+  'abcd-extended-version': abcdExtendedVersion,
+  'beyond-end-credit-version': beyondEndCreditVersion,
+  'kimagure-romantic': kimagureRomantic,
+  'era-with-nayeon': eraWithNayeon,
+  'twenty-five-twenty-one': twentyFiveTwentyOne,
+  zone: zone,
+  'killin-me-good-english': killinMeGoodEnglishVer,
+  'marry-my-husband-ost-part2': marryMyHusbandOstPart2,
+  'abou-tzu': abouTzu,
+  blink: blink,
+  'lil-fantasy-vol1': lilFantasyVol1,
+  bouquet: bouquet,
+  masterpiece: masterpiece,
+  'haute-couture': hauteCouture,
+  message: message,
+  play: play,
+}
+
+export const DISCOGRAPHY: TDiscography[] = DISCOGRAPHY_DATA.map((data) => ({
+  ...data,
+  image: ALBUM_IMAGES[data.id],
+}))
