+++ src/const/members-data.ts
@@ -1,27 +1,21 @@
-// ⚠️  IMPORTANT: Remember to sync src/tests/const/members.ts with this file
-// When updating member data, copy the changes here but exclude image properties
+import type { TMemberData } from '../types/members'
+import type { TUnit } from '../types/discography'
 
-import type { TMember } from '../../types/members'
-import nayeon from '../../assets/members/cards/nayeon.webp'
-import jeongyeon from '../../assets/members/cards/jeongyeon.webp'
-import momo from '../../assets/members/cards/momo.webp'
-import sana from '../../assets/members/cards/sana.webp'
-import jihyo from '../../assets/members/cards/jihyo.webp'
-import mina from '../../assets/members/cards/mina.webp'
-import dahyun from '../../assets/members/cards/dahyun.webp'
-import chaeyoung from '../../assets/members/cards/chaeyoung.webp'
-import tzuyu from '../../assets/members/cards/tzuyu.webp'
-import nayeonFull from '../../assets/members/full/nayeon-full.webp'
-import jeongyeonFull from '../../assets/members/full/jeongyeon-full.webp'
-import momoFull from '../../assets/members/full/momo-full.webp'
-import sanaFull from '../../assets/members/full/sana-full.webp'
-import jihyoFull from '../../assets/members/full/jihyo-full.webp'
-import minaFull from '../../assets/members/full/mina-full.webp'
-import dahyunFull from '../../assets/members/full/dahyun-full.webp'
-import chaeyoungFull from '../../assets/members/full/chaeyoung-full.webp'
-import tzuyuFull from '../../assets/members/full/tzuyu-full.webp'
+export const Unit = {
+  twice: 'twice',
+  misamo: 'misamo',
+  nayeon: 'nayeon',
+  jeongyeon: 'jeongyeon',
+  momo: 'momo',
+  sana: 'sana',
+  jihyo: 'jihyo',
+  mina: 'mina',
+  dahyun: 'dahyun',
+  chaeyoung: 'chaeyoung',
+  tzuyu: 'tzuyu',
+} as const satisfies Record<string, TUnit>
 
-export const MEMBERS: readonly TMember[] = [
+export const MEMBERS_DATA: readonly TMemberData[] = [
   {
     id: 'nayeon',
     name: 'Nayeon',
@@ -38,8 +32,6 @@ export const MEMBERS: readonly TMember[] = [
     nationality: 'South Korean',
     birthday: new Date('1995-09-22'),
     positions: ['Main Vocalist', 'Center', 'Group Face'],
-    image: nayeon,
-    fullImage: nayeonFull,
     songs: [
       {
         title: 'POP!',
@@ -155,8 +147,6 @@ export const MEMBERS: readonly TMember[] = [
     nationality: 'South Korean',
     birthday: new Date('1996-11-01'),
     positions: ['Main Vocalist'],
-    image: jeongyeon,
-    fullImage: jeongyeonFull,
     songs: [
       {
         title: 'FIX A DRINK',
@@ -247,8 +237,6 @@ export const MEMBERS: readonly TMember[] = [
     nationality: 'Japanese',
     birthday: new Date('1996-11-09'),
     positions: ['Main Dancer', 'Sub-Vocalist', 'Sub-Rapper'],
-    image: momo,
-    fullImage: momoFull,
     songs: [
       {
         title: 'MOVE LIKE THAT',
@@ -364,8 +352,6 @@ export const MEMBERS: readonly TMember[] = [
     nationality: 'Japanese',
     birthday: new Date('1996-12-29'),
     positions: ['Sub-Vocalist'],
-    image: sana,
-    fullImage: sanaFull,
     songs: [
       {
         title: 'DECAFFEINATED',
@@ -481,8 +467,6 @@ export const MEMBERS: readonly TMember[] = [
     nationality: 'South Korean',
     birthday: new Date('1997-02-01'),
     positions: ['Leader', 'Main Vocalist'],
-    image: jihyo,
-    fullImage: jihyoFull,
     songs: [
       {
         title: "Killin' Me Good",
@@ -594,8 +578,6 @@ export const MEMBERS: readonly TMember[] = [
     nationality: 'Japanese',
     birthday: new Date('1997-03-24'),
     positions: ['Main Dancer', 'Sub-Vocalist'],
-    image: mina,
-    fullImage: minaFull,
     songs: [
       {
         title: 'STONE COLD',
@@ -711,8 +693,6 @@ export const MEMBERS: readonly TMember[] = [
     nationality: 'South Korean',
     birthday: new Date('1997-05-28'),
     positions: ['Sub-Vocalist', 'Main-Rapper'],
-    image: dahyun,
-    fullImage: dahyunFull,
     songs: [
       {
         title: 'CHESS',
@@ -808,8 +788,6 @@ export const MEMBERS: readonly TMember[] = [
     nationality: 'South Korean',
     birthday: new Date('1997-04-23'),
     positions: ['Sub-Vocalist', 'Main-Rapper'],
-    image: chaeyoung,
-    fullImage: chaeyoungFull,
     songs: [
       {
         title: 'SHOOT (Firecracker)',
@@ -937,8 +915,6 @@ export const MEMBERS: readonly TMember[] = [
     nationality: 'Taiwanese',
     birthday: new Date('1999-06-14'),
     positions: ['Sub-Vocalist', 'Visual', 'Maknae'],
-    image: tzuyu,
-    fullImage: tzuyuFull,
     songs: [
       {
         title: 'Run Away',
