+++ src/const/members.ts
@@ -1,23 +1,5 @@
-// ⚠️  IMPORTANT: Remember to sync src/tests/const/members.ts with this file
-// When updating member data, copy the changes here but exclude image properties
-
-import type { TMember } from '../types/members'
-import type { TUnit } from '../types/discography'
-
-export const Unit = {
-  twice: 'twice',
-  misamo: 'misamo',
-  nayeon: 'nayeon',
-  jeongyeon: 'jeongyeon',
-  momo: 'momo',
-  sana: 'sana',
-  jihyo: 'jihyo',
-  mina: 'mina',
-  dahyun: 'dahyun',
-  chaeyoung: 'chaeyoung',
-  tzuyu: 'tzuyu',
-} as const satisfies Record<string, TUnit>
-
+import type { TMember, TMemberId } from '../types/members'
+import { MEMBERS_DATA, Unit } from './members-data'
 import nayeon from '../assets/members/cards/nayeon.webp'
 import jeongyeon from '../assets/members/cards/jeongyeon.webp'
 import momo from '../assets/members/cards/momo.webp'
@@ -37,1009 +19,21 @@ import dahyunFull from '../assets/members/full/dahyun-full.webp'
 import chaeyoungFull from '../assets/members/full/chaeyoung-full.webp'
 import tzuyuFull from '../assets/members/full/tzuyu-full.webp'
 
-export const MEMBERS: readonly TMember[] = [
-  {
-    id: 'nayeon',
-    name: 'Nayeon',
-    fullName: 'Im Nayeon',
-    emoji: '🐰',
-    social: [
-      {
-        platform: 'instagram',
-        username: 'nayeonyny',
-        url: 'https://www.instagram.com/nayeonyny/',
-      },
-    ],
-    color: 'sky blue',
-    nationality: 'South Korean',
-    birthday: new Date('1995-09-22'),
-    positions: ['Main Vocalist', 'Center', 'Group Face'],
-    image: nayeon,
-    fullImage: nayeonFull,
-    songs: [
-      {
-        title: 'POP!',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/f6YDKF0LVWw?si=lVo3cXGAHFhGdDMF',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=R9TPed_ohKM&si=LCsSQtWAqSQyABtQ',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/3lOMJTQTd6J34faYwASc33?si=7d2fc6f03dee4e03',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/pop/1624932774',
-          },
-        ],
-      },
-      {
-        title: 'ABCD',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/oUZttxRcPZw?list=RDoUZttxRcPZw',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=4G30iKdqGeg&si=8PjzzUAH2x75L1IK',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/0V2passWyAXnON67kfAj7y?si=56e0069a56334793',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/abcd/1745594619',
-          },
-        ],
-      },
-      {
-        title: 'MEEEEEE',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/16_2Z--YW7s?list=RD16_2Z--YW7s',
-          },
-        ],
-      },
-      {
-        title: 'LOVE COUNTDOWN',
-        links: [
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=D6KP4btz2eY&si=XOz33z4XYseWh6H7',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/0GnydTKmczMoKOEl1D4URI?si=b2c5e6f03d9c4044',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/love-countdown-feat-wonstein/1624932938',
-          },
-        ],
-      },
-      {
-        title: 'NO PROBLEM',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/XRQxpx4NGUk?list=RDXRQxpx4NGUk',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=9oelHREj8iI&si=UIx3jIro9dfd2urd',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/4zHvWi4iFAG45lgiN7smLC?si=97468ebe982741c1',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/no-problem-feat-felix/1624932781',
-          },
-        ],
-      },
-    ],
-    funFacts: [
-      'Her height is 163cm',
-      'Her weight is 47kg',
-      'She has a younger sister, Seoyeon, born 1998.',
-      'She chose 9 as her favorite number, because TWICE has 9 members',
-      'Her left leg is weak due to her being in a car accident in her childhood.',
-    ],
-  },
-  {
-    id: 'jeongyeon',
-    name: 'Jeongyeon',
-    fullName: 'Yoo Jeong-yeon',
-    emoji: '🐶',
-    social: [
-      {
-        platform: 'instagram',
-        username: 'jy_piece',
-        url: 'https://www.instagram.com/jy_piece/',
-      },
-    ],
-    color: 'green',
-    nationality: 'South Korean',
-    birthday: new Date('1996-11-01'),
-    positions: ['Main Vocalist'],
-    image: jeongyeon,
-    fullImage: jeongyeonFull,
-    songs: [
-      {
-        title: 'FIX A DRINK',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/2hMplq1eXOo?list=RD2hMplq1eXOo',
-          },
-        ],
-      },
-      {
-        title: 'TAKEDOWN',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/vf1kWWqJKus?list=RDvf1kWWqJKus',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=7XRcflf_E0c&si=W86IU3El65Vij1f8',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/1rKQjUhF9zFJmuUotr3VkV?si=9c96b5dc002d4b6a',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/takedown-jeongyeon-jihyo-chaeyoung/1820264142',
-          },
-        ],
-      },
-      {
-        title: 'BATTITUDE',
-        links: [
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=TOic7n-CDd4&si=jz5kz4HN731LZSAR',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/4AdIi71e7MCdNwy4T9VfTW?si=b090e2f890d249de',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/battitude-nayeon-jeongyeon-momo-mina/1813491330',
-          },
-        ],
-      },
-      {
-        title: 'LET LOVE GO',
-        links: [
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=dTsP-nmggyc&si=jx0An2eCUMMcbPbJ',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/0v6BSRyLyo6fHBlPop5EvR?si=d9d0333470584cea',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/let-love-go-jeongyeon-momo-sana-tzuyu/1813491339',
-          },
-        ],
-      },
-    ],
-    funFacts: [
-      'Her height is 167cm',
-      'Her weight is 50kg',
-      'She has 2 older sisters: Gong Seungyeon (actress) and Seoyeon (office worker)',
-      "She chose 0 as her favorite number, because she doesn't actually have a favorite number",
-      'Her father, Yoo Changjoon, is a popular Korean traditional chef',
-    ],
-  },
-  {
-    id: 'momo',
-    name: 'Momo',
-    fullName: 'Hirai Momo',
-    emoji: '🍑',
-    social: [
-      {
-        platform: 'instagram',
-        username: 'momo',
-        url: 'https://www.instagram.com/momo/',
-      },
-    ],
-    color: 'pink',
-    nationality: 'Japanese',
-    birthday: new Date('1996-11-09'),
-    positions: ['Main Dancer', 'Sub-Vocalist', 'Sub-Rapper'],
-    image: momo,
-    fullImage: momoFull,
-    songs: [
-      {
-        title: 'MOVE LIKE THAT',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/mfDt-GWZdFc?si=D-63NX_qEGF5mHE6',
-          },
-        ],
-      },
-      {
-        title: 'Money In My Pocket',
-        links: [
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=NFAi_mxlcXs&si=RCKLxPhKevPPxAiU',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/78qBcW44wNzdr2EDpqi0oV?si=0a98abac75e24b4a',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/money-in-my-pocket-momo/1772137194',
-          },
-        ],
-      },
-      {
-        title: 'Do not touch',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/woKq2sD8xho?si=x7kks_wfSKmpz7oZ',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=w9VHb4BtfYY&si=momrXQCe3blVNKnp',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/1w6b63TmaXFk1jhPOY0FkY?si=8defd9cbe6324c93',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/do-not-touch/1691852612',
-          },
-        ],
-      },
-      {
-        title: 'Identity',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/-J_w9K6DgRo?si=tJvv9wmncqHdVjEs',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=d67H9wtS7rs&si=abmMdMjlyDFqo0XW',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/0wu7L2JkRqx2nFCSiYqcKy?si=13cec2a95ead43e7',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/identity/1772136985',
-          },
-        ],
-      },
-      {
-        title: 'Marshmallow',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/gAFgaI5D3Gc?si=3tGXjSPQ3IPWcPqF',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=TPUG4dDuG_I&si=stX1TY5xkfHZeBgS',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/1V6H3YoNZJ2WlPmwsaOXZx?si=b3cf8455afc24ca9',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/marshmallow/1691852623',
-          },
-        ],
-      },
-    ],
-    funFacts: [
-      'Her height is 167cm',
-      'Her weight is 48kg',
-      'She has an older sister, Hana, born in 1994.',
-      "She chose 64 as her favorite number, because it's the number on her father's shirt when he plays football",
-      'She was scouted by JYP Entertainment after they saw a dance video of her and her older sister.',
-    ],
-  },
-  {
-    id: 'sana',
-    name: 'Sana',
-    fullName: 'Minatozaki Sana',
-    emoji: '🐭',
-    social: [
-      {
-        platform: 'instagram',
-        username: 'm.by__sana',
-        url: 'https://www.instagram.com/m.by__sana/',
-      },
-    ],
-    color: 'purple',
-    nationality: 'Japanese',
-    birthday: new Date('1996-12-29'),
-    positions: ['Sub-Vocalist'],
-    image: sana,
-    fullImage: sanaFull,
-    songs: [
-      {
-        title: 'DECAFFEINATED',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/wfsrUojDPtg?si=82GzSbDYw01LUTEu',
-          },
-        ],
-      },
-      {
-        title: 'Mirage',
-        links: [
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=KeVK8bQzhPc&si=GaJPPrlyQXtgj--m',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/0HtVZMMehKHJPG4npCnXIv?si=80db681ba1b04739',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/mirage-sana/1772137195',
-          },
-        ],
-      },
-      {
-        title: 'Do not touch',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/woKq2sD8xho?si=x7kks_wfSKmpz7oZ',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=w9VHb4BtfYY&si=momrXQCe3blVNKnp',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/1w6b63TmaXFk1jhPOY0FkY?si=8defd9cbe6324c93',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/do-not-touch/1691852612',
-          },
-        ],
-      },
-      {
-        title: 'Identity',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/-J_w9K6DgRo?si=tJvv9wmncqHdVjEs',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=d67H9wtS7rs&si=abmMdMjlyDFqo0XW',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/0wu7L2JkRqx2nFCSiYqcKy?si=13cec2a95ead43e7',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/identity/1772136985',
-          },
-        ],
-      },
-      {
-        title: 'Marshmallow',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/gAFgaI5D3Gc?si=3tGXjSPQ3IPWcPqF',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=TPUG4dDuG_I&si=stX1TY5xkfHZeBgS',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/1V6H3YoNZJ2WlPmwsaOXZx?si=b3cf8455afc24ca9',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/marshmallow/1691852623',
-          },
-        ],
-      },
-    ],
-    funFacts: [
-      'Her height is 168cm',
-      'Her weight is 47kg',
-      'She is an only child.',
-      'She chose 12 as her favorite number (her birthday month), because Mina took the number 37.',
-      'She likes spicy foods.',
-    ],
-  },
-  {
-    id: 'jihyo',
-    name: 'Jihyo',
-    fullName: 'Park Ji-hyo',
-    emoji: '🦄',
-    social: [
-      {
-        platform: 'instagram',
-        username: '_zyozyo',
-        url: 'https://www.instagram.com/_zyozyo/',
-      },
-    ],
-    color: 'apricot',
-    nationality: 'South Korean',
-    birthday: new Date('1997-02-01'),
-    positions: ['Leader', 'Main Vocalist'],
-    image: jihyo,
-    fullImage: jihyoFull,
-    songs: [
-      {
-        title: "Killin' Me Good",
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/jyw_vrI4ySg?si=HDvlvXEH3tDQsK5a',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=bJ3WGDFznis&si=QE6JrkxE0w00Gl9X',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/4Cyf87ul1le0xTWRFBoYd7?si=be81c51f761e482a',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/killin-me-good/1697360132',
-          },
-        ],
-      },
-      {
-        title: 'Closer',
-        links: [
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=71c1dJBL-DU&si=-mtGx2ISTuJ4z53g',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/0aUNIY0nITYfGH1ooJBRvR?si=bfaf705415bf4382',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/closer/1697360323',
-          },
-        ],
-      },
-      {
-        title: 'ATM',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/lY8jvTG0edQ?si=zzl22wWyMgGFTClV',
-          },
-        ],
-      },
-      {
-        title: 'Talkin About It',
-        links: [
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=mFyAItymIBg&si=66k2_kTWOrikMnAw',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/28mFFxAwJ9cx9j7z9yu64R?si=8f1c33633d8844f7',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/talkin-about-it/1697360300',
-          },
-        ],
-      },
-      {
-        title: 'TAKEDOWN',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/vf1kWWqJKus?list=RDvf1kWWqJKus',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=7XRcflf_E0c&si=W86IU3El65Vij1f8',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/1rKQjUhF9zFJmuUotr3VkV?si=9c96b5dc002d4b6a',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/takedown-jeongyeon-jihyo-chaeyoung/1820264142',
-          },
-        ],
-      },
-    ],
-    funFacts: [
-      'Her height is 162cm',
-      'Her weight is 56kg',
-      'She has 2 younger sisters, Seoyeon (born in 2008)  and Jiyoung (born in 2002)',
-      'She chose 21 as her favorite number, because she was born on February 1st.',
-      'She was voted leader by the members, not JYPE (by anonymous voting).',
-    ],
-  },
-  {
-    id: 'mina',
-    name: 'Mina',
-    fullName: 'Myoui Mina',
-    emoji: '🐧',
-    social: [
-      {
-        platform: 'instagram',
-        username: 'mina_sr_my',
-        url: 'https://www.instagram.com/mina_sr_my/',
-      },
-    ],
-    color: 'mint',
-    nationality: 'Japanese',
-    birthday: new Date('1997-03-24'),
-    positions: ['Main Dancer', 'Sub-Vocalist'],
-    image: mina,
-    fullImage: minaFull,
-    songs: [
-      {
-        title: 'STONE COLD',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/XXwo3aOVLxU?si=OJGpqk9-LDkan1WH',
-          },
-        ],
-      },
-      {
-        title: 'Misty',
-        links: [
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=wkXE4-VpSec&si=zrua6tW18xbzCRGT',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/347F32RMBGXtUyntcqX1oS?si=7fb12800d30d4970',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/misty-mina/1772137192',
-          },
-        ],
-      },
-      {
-        title: 'Do not touch',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/woKq2sD8xho?si=x7kks_wfSKmpz7oZ',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=w9VHb4BtfYY&si=momrXQCe3blVNKnp',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/1w6b63TmaXFk1jhPOY0FkY?si=8defd9cbe6324c93',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/do-not-touch/1691852612',
-          },
-        ],
-      },
-      {
-        title: 'Identity',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/-J_w9K6DgRo?si=tJvv9wmncqHdVjEs',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=d67H9wtS7rs&si=abmMdMjlyDFqo0XW',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/0wu7L2JkRqx2nFCSiYqcKy?si=13cec2a95ead43e7',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/identity/1772136985',
-          },
-        ],
-      },
-      {
-        title: 'Marshmallow',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/gAFgaI5D3Gc?si=3tGXjSPQ3IPWcPqF',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=TPUG4dDuG_I&si=stX1TY5xkfHZeBgS',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/1V6H3YoNZJ2WlPmwsaOXZx?si=b3cf8455afc24ca9',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/marshmallow/1691852623',
-          },
-        ],
-      },
-    ],
-    funFacts: [
-      'Her height is 163cm',
-      'Her weight is 46kg',
-      'She has an older brother, Kai, born in 1992.',
-      'She was born in San Antonio, Texas, USA.',
-      "She chose 37 as her favorite number, because '37' in Japanese it's pronounced as 'Mi' or 'San' (3) and 'Nana' [often shortened to Na] (7)",
-    ],
-  },
-  {
-    id: 'dahyun',
-    name: 'Dahyun',
-    fullName: 'Kim Da-hyun',
-    emoji: '💛',
-    social: [
-      {
-        platform: 'instagram',
-        username: 'dahhyunnee',
-        url: 'https://www.instagram.com/dahhyunnee/',
-      },
-    ],
-    color: 'white',
-    nationality: 'South Korean',
-    birthday: new Date('1997-05-28'),
-    positions: ['Sub-Vocalist', 'Main-Rapper'],
-    image: dahyun,
-    fullImage: dahyunFull,
-    songs: [
-      {
-        title: 'CHESS',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/KGIEoc7Ry5k?si=KqApRKgF2n1qlNNp',
-          },
-        ],
-      },
-      {
-        title: 'PUSH & PULL',
-        links: [
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=LKHm2N-knDQ&si=N0ey4e8BXP6dG8ns',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/7rY4s8axCW9tQXboPQ46nJ?si=52f13b39c69f42c1',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/push-pull/1591510033',
-          },
-        ],
-      },
-      {
-        title: 'Switch to me',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/5UMgw3V-Q5I?si=_DmMI5aY94wTKGVF',
-          },
-        ],
-      },
-      {
-        title: 'G.O.A.T.',
-        links: [
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=nJRj6-jeSNE&si=zQ3WCYntdXWAU0X5',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/4Mp85pG3BuRG5ga6wDyvHA?si=85dbde1ed04c4c65',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/g-o-a-t-mina-dahyun-chaeyoung/1813491340',
-          },
-        ],
-      },
-      {
-        title: 'DAT AHH DAT OOH',
-        links: [
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=POxNm2NNyps&si=EtsbQWcr2_ZoVwiU',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/7JL2YMT3SAjVuLp0znxT4a?si=0b3147dc877a478c',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/dat-ahh-dat-ooh-sana-jihyo-dahyun-chaeyoung-tzuyu/1813491335',
-          },
-        ],
-      },
-    ],
-    funFacts: [
-      'Her height is 165cm',
-      'Her weight is 48kg',
-      'She has an older brother, Myungsoo.',
-      "She chose 7 as her favorite number, because she became a TWICE member on July 7 and because she's the 7th oldest in TWICE.",
-      'She is scared of animals (cats, dogs, etc.).',
-    ],
-  },
-  {
-    id: 'chaeyoung',
-    name: 'Chaeyoung',
-    fullName: 'Son Chae-young',
-    emoji: '🍓',
-    social: [
-      {
-        platform: 'instagram',
-        username: 'chaeyo.0',
-        url: 'https://www.instagram.com/chaeyo.0/',
-      },
-    ],
-    color: 'red',
-    nationality: 'South Korean',
-    birthday: new Date('1997-04-23'),
-    positions: ['Sub-Vocalist', 'Main-Rapper'],
-    image: chaeyoung,
-    fullImage: chaeyoungFull,
-    songs: [
-      {
-        title: 'SHOOT (Firecracker)',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/tl2luF-JRlU?si=eSEy3zALK9es9-iw',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=glptno5ohjc&si=W2h1L7jzUQ1L7Lzs',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/0pUEewaz1L9Vt8BCEVEnxx?si=2be5c600d21d452c',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/shoot-firecracker/1832584204',
-          },
-        ],
-      },
-      {
-        title: 'AVOCADO',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/NlHD4ndO-z0?si=1SRbpmbEWda-2vYr',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=SGwinvCq4oY&si=Pltn03fSMiMnzMxx',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/1fYSCk0DhxprTVb1B2zg4Q?si=3e2947e95e044bf4',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/avocado-feat-gliiico/1832584200',
-          },
-        ],
-      },
-      {
-        title: 'GIRL',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/ML3knJPwNTA?si=UFSgsWBMs5UbEd0U',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=_pM7B1cej_8&si=tV9aVZH_bWEvf_Uq',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/5yIO3LTR0BcDjYxliK9hFV?si=41b8b8337da04f79',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/girl/1832584207',
-          },
-        ],
-      },
-      {
-        title: 'TAKEDOWN',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/vf1kWWqJKus?list=RDvf1kWWqJKus',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=7XRcflf_E0c&si=W86IU3El65Vij1f8',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/1rKQjUhF9zFJmuUotr3VkV?si=9c96b5dc002d4b6a',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/takedown-jeongyeon-jihyo-chaeyoung/1820264142',
-          },
-        ],
-      },
-      {
-        title: '그림자놀이',
-        links: [
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=jqNpkGKf0Cw&si=0Dk9pYN3v3Inuu0u',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/7kwFRE8IHukpi9UiULk3oL?si=d56d0c8a668441c0',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/%EA%B7%B8%EB%A6%BC%EC%9E%90%EB%86%80%EC%9D%B4/1832584699',
-          },
-        ],
-      },
-    ],
-    funFacts: [
-      'Her height is 163cm',
-      'Her weight is 46kg',
-      'She has a younger brother, Jeonghun, born in 2000, who is an aspiring model.',
-      'She chose 29 as her favorite number, because 2 (TWICE) and 9 (number of members)',
-      'On April 5, 2024, it was rumored that she is dating Korean rapper Zion.T before both of their companies confirmed the rumor on the same day.',
-    ],
-  },
-  {
-    id: 'tzuyu',
-    name: 'Tzuyu',
-    fullName: 'Chou Tzuyu',
-    emoji: '🦌',
-    social: [
-      {
-        platform: 'instagram',
-        username: 'thinkaboutzu',
-        url: 'https://www.instagram.com/thinkaboutzu/',
-      },
-    ],
-    color: 'blue',
-    nationality: 'Taiwanese',
-    birthday: new Date('1999-06-14'),
-    positions: ['Sub-Vocalist', 'Visual', 'Maknae'],
-    image: tzuyu,
-    fullImage: tzuyuFull,
-    songs: [
-      {
-        title: 'Run Away',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/IIrCrGAX03Y?si=jJ8Pqb597VZ8Qje8',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=YofelpFMWzg&si=DzLDrfqqpBquxwGH',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/2Ti4DCTo6Ag7QvaYex8A8B?si=b6c5273fcf004ac1',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/run-away/1760441617',
-          },
-        ],
-      },
-      {
-        title: 'DIVE IN',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/jI_rZqZcSEA?si=qzmiVKqjCzrL9_ox',
-          },
-        ],
-      },
-      {
-        title: 'Heartbreak In Heaven',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/EV-SV9YzRdQ?si=sn-o0hegXla5TAsu',
-          },
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=gLtEq5UtPvk&si=NmxNwxsbqlcEwZkI',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/4EVa8hqOEfg2NZiruvgLnQ?si=1b7894eb8e3b4348',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/heartbreak-in-heaven-feat-peniel/1760441622',
-          },
-        ],
-      },
-      {
-        title: 'ME!',
-        links: [
-          {
-            platform: 'youtube',
-            url: 'https://youtu.be/IO1Nl9ErTsI?si=Mkv3J5GszQ_7vgS7',
-          },
-        ],
-      },
-      {
-        title: 'LET LOVE GO',
-        links: [
-          {
-            platform: 'youtube-music',
-            url: 'https://music.youtube.com/watch?v=dTsP-nmggyc&si=jx0An2eCUMMcbPbJ',
-          },
-          {
-            platform: 'spotify',
-            url: 'https://open.spotify.com/track/0v6BSRyLyo6fHBlPop5EvR?si=d9d0333470584cea',
-          },
-          {
-            platform: 'appleMusic',
-            url: 'https://music.apple.com/us/song/let-love-go-jeongyeon-momo-sana-tzuyu/1813491339',
-          },
-        ],
-      },
-    ],
-    funFacts: [
-      'Her height is 170cm',
-      'Her weight is 48kg',
-      'She has an older brother, born in 1992.',
-      'Her favorite number is 25, because she thinks it is a cool number.',
-      'She is the tallest member in the group.',
-    ],
-  },
-] as const
+export { Unit }
+
+const MEMBER_IMAGES: Record<TMemberId, { image: ImageMetadata; fullImage: ImageMetadata }> = {
+  nayeon: { image: nayeon, fullImage: nayeonFull },
+  jeongyeon: { image: jeongyeon, fullImage: jeongyeonFull },
+  momo: { image: momo, fullImage: momoFull },
+  sana: { image: sana, fullImage: sanaFull },
+  jihyo: { image: jihyo, fullImage: jihyoFull },
+  mina: { image: mina, fullImage: minaFull },
+  dahyun: { image: dahyun, fullImage: dahyunFull },
+  chaeyoung: { image: chaeyoung, fullImage: chaeyoungFull },
+  tzuyu: { image: tzuyu, fullImage: tzuyuFull },
+}
+
+export const MEMBERS: readonly TMember[] = MEMBERS_DATA.map((data) => ({
+  ...data,
+  ...MEMBER_IMAGES[data.id],
+}))
